requirements.txt
python-dotenv
django-crispy-forms
djangorestframework
django-cors-headers
django-recaptcha==2.0.6